move TapPressButton into separate library files
