+++ lib/TapPressButton/TapPressButton.cpp
@@ -1,0 +1,14 @@
+#include "TapPressButton.h"
+
+    void TapPressButton::updateInput(bool input) {
+        prevButtonState = currentButtonState;
+        currentButtonState = input;
+    }
+    
+    bool TapPressButton::getCurrentState() {
+        return currentButtonState;
+    }
+    
+    bool TapPressButton::hasChanged() {
+        return currentButtonState != prevButtonState;
+    }
